[genius] Adding normalization options to GeniusNegotiator
The normalize_utility and normalize_max_only options will apply to
any ufun set for the negotiator up to the beginning of the negotiation.
They will not apply to assignments of ufun after the beginning of the
negotiation (which should not happen anyway). It does not matter
whether the source of the ufun is a file or a ufun passed to the
negotiator.