[testing] avoiding failure on genius agents that cannot agree
Some genius agents (AGENTS_WITH_NO_AGREEMENT_ON_SAME_UFUN in
test_genius.py) cannot get an agreement when negotiating with an
opponent using the same ufun as the opponent!!
This is strange but not strictly forbidden behavior. I am keeping
a list of these agents for future testing.
TODO revisit this issue