[helpers] allowing users to force single thread
This is of value for debugging. If force_single_thread() is called (with
True which is the default), NegMAS will not create a separate thead for
each call to an agent or negotiator which imply not enforcing
time-limits except after the fact (i.e. after the return of the call).