Compatibility with genius-bridge v0.11
When a GeniusNegotiator creates a bridge (i.e. when one is started with
no bridge known to the system), it will start it so that nothing gets
printed or logged. This is both more efficient and elegant as it avoids
dumping nonnsense to the screen. We cannot though avoid printing from
Genius agents themselves.