making ContiguousIssue and AffineFun save discrete
An issue when dealing with genius negotiators is that some of them try
to explicitly conver the issue into an IssueDiscrete which cannot be
done if the issue itself was IssueInteger.

Originally, we used an efficient way to save issue values and linear
ufun values but it was leading to some negotiators failing because they
expect discrete valued issues and integer issues are not considered
discrete. I do not have any idea why.

As a result, we are forced to use a much less efficient way to save
issues and linear ufuns to xml for compatibility with Genius.